+++ mne/viz/_3d.py
@@ -40,8 +40,8 @@ from ..transforms import (read_trans, _find_trans, apply_trans, rot_to_quat,
 from ..utils import (get_subjects_dir, logger, _check_subject, verbose, warn,
                      _import_mlab, SilenceStdout, has_nibabel, check_version,
                      _ensure_int, _validate_type)
-from .utils import (mne_analyze_colormap, _prepare_trellis, _get_color_list,
-                    plt_show, tight_layout, figure_nobar, _check_time_unit)
+from .utils import (mne_analyze_colormap, _prepare_trellis, COLORS, plt_show,
+                    tight_layout, figure_nobar, _check_time_unit)
 from ..bem import (ConductorModel, _bem_find_surface, _surf_dict, _surf_name,
                    read_bem_surfaces)
 
@@ -536,7 +536,6 @@ def _plot_mri_contours(mri_fname, surf_fnames, orientation='coronal',
         # and then plot the contours on top
         for surf in surfs:
             with warnings.catch_warnings(record=True):  # no contours
-                warnings.simplefilter('ignore')
                 ax.tricontour(surf['rr'][:, inds[0]], surf['rr'][:, inds[1]],
                               surf['tris'], surf['rr'][:, inds[2]],
                               levels=[sl], colors='yellow', linewidths=2.0)
@@ -1344,7 +1343,7 @@ def _key_pressed_slider(event, params):
 
 def _smooth_plot(this_time, params):
     """Smooth source estimate data and plot with mpl."""
-    from ..source_estimate import _morph_buffer
+    from ..morph import _morph_buffer
     from mpl_toolkits.mplot3d import art3d
     ax = params['ax']
     stc = params['stc']
@@ -1404,7 +1403,7 @@ def _plot_mpl_stc(stc, subject=None, surface='inflated', hemi='lh',
     from matplotlib.widgets import Slider
     import nibabel as nib
     from scipy import sparse, stats
-    from ..source_estimate import _get_subject_sphere_tris
+    from ..morph import _get_subject_sphere_tris
     if hemi not in ['lh', 'rh']:
         raise ValueError("hemi must be 'lh' or 'rh' when using matplotlib. "
                          "Got %s." % hemi)
@@ -1971,7 +1970,7 @@ def plot_sparse_source_estimates(src, stcs, colors=None, linewidth=2,
         labels = [labels]
 
     if colors is None:
-        colors = _get_color_list()
+        colors = COLORS
 
     linestyles = ['-', '--', ':']